requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=3.10,<4.0
Pillow>=11.2
django-allauth>=0.45,<1.0
django-filter>=2.4,<3.0
django-ckeditor>=6.0,<7.0
django-storages>=1.10,<2.0
boto3>=1.18,<2.0